feat(): Translation settings + user configuration
