+++ packages/modules/translation/src/utils/filter-translation-fields.ts
@@ -1,0 +1,27 @@
+import { TranslationTypes } from "@medusajs/framework/types"
+
+export function filterTranslationFields(
+  translations: TranslationTypes.TranslationDTO[],
+  translatableFieldsConfig: Record<string, string[]>
+): TranslationTypes.TranslationDTO[] {
+  return translations.map((translation) => {
+    const allowedFields = translatableFieldsConfig[translation.reference]
+    if (!allowedFields || allowedFields.length === 0) {
+      return { ...translation, translations: {} }
+    }
+
+    const filteredTranslations: Record<string, unknown> = {}
+    for (const field of allowedFields) {
+      if (
+        translation.translations &&
+        field in (translation.translations as Record<string, unknown>)
+      ) {
+        filteredTranslations[field] = (
+          translation.translations as Record<string, unknown>
+        )[field]
+      }
+    }
+
+    return { ...translation, translations: filteredTranslations }
+  })
+}
